Add files via upload
This file contains the results (Excel files) of integer model with 9 variables (referred to as model 1), integer model with 5 variables (referred to as model 2), and fractional model with 5 variables (referred to as model 3) corresponding to the states of the Group of Twenty with predictions of the future evolution of the GDP (analysis for N = 5, 10 and 15 years). 

These are the results for the paper entitled "Fractional Derivatives for Economic Growth Modelling of the Group of Twenty: Application to Prediction" by Inés Tejado, Emiliano Pérez, Duarte Valério submitted to Mathematics journal (November 2019).